pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sf-rotation"
version = "1.0.0"
description = "Snowflake Key Pair Rotation Tool for Hevo Data destinations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["snowflake", "hevo", "key-pair", "rotation", "authentication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Security",
]
dependencies = [
    "snowflake-connector-python>=3.0.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
sf-rotation = "sf_rotation.main:main"

[project.urls]
Homepage = "https://github.com/Legolasan/sf_rotation"
Repository = "https://github.com/Legolasan/sf_rotation"
Issues = "https://github.com/Legolasan/sf_rotation/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sf_rotation = ["*.yaml", "*.yml"]
